+++ src/helpers.js
@@ -9,7 +9,8 @@ const path = require('path');
  * @property {String} destinationPath
  *  - Path to output destination (default 'dist', changes to 'debug' during dev-server)
  * @property {Object} globalMeta - Meta variables that are accessible globally in .abell files
- *
+ * @property {Array} ignoreInBuild
+ * 
  */
 
 const relativeJoinedPath = (pathString) =>
@@ -74,20 +75,28 @@ const getAbellFiles = (sourcePath, extension) => {
  */
 function getAbellConfigs() {
   let abellConfig;
+  const defaultConfigs = {
+    destinationPath: 'dist',
+    sourcePath: 'theme',
+    contentPath: 'content',
+    ignoreInBuild: [],
+    globalMeta: {}
+  };
+
   try {
+    // In dev-server, user may change the configs so in that case we should drop the old cache
     delete require.cache[path.join(process.cwd(), 'abell.config.js')];
-    abellConfig = require(path.join(process.cwd(), 'abell.config.js'));
+    abellConfig = {
+      ...defaultConfigs,
+      ...require(path.join(process.cwd(), 'abell.config.js'))
+    };
+
     if (Object.keys(abellConfig).length <= 0) {
       throw new Error( `Something went wrong while fetching new configurations. Save again to refresh the dev server.` ); // eslint-disable-line
     }
   } catch (err) {
     console.log(boldRed('>> ') + err.message);
-    abellConfig = {
-      destinationPath: 'dist',
-      sourcePath: 'theme',
-      contentPath: 'content',
-      globalMeta: {},
-    };
+    abellConfig = defaultConfigs;
   }
 
   const destinationPath = relativeJoinedPath(abellConfig.destinationPath);
@@ -112,15 +121,19 @@ const createPathIfAbsent = (pathToCreate) => {
  *
  * @param {String} from - Path to copy from
  * @param {String} to - Path to copy to
+ * @param {Array} ignore - files/directories to ignore
  * @return {void}
  */
-function copyFolderSync(from, to) {
+function copyFolderSync(from, to, ignore = []) {
+  if (ignore.includes(from)) {
+    return;
+  };
   createPathIfAbsent(to);
   fs.readdirSync(from).forEach((element) => {
     if (fs.lstatSync(path.join(from, element)).isFile()) {
       fs.copyFileSync(path.join(from, element), path.join(to, element));
     } else {
-      copyFolderSync(path.join(from, element), path.join(to, element));
+      copyFolderSync(path.join(from, element), path.join(to, element), ignore);
     }
   });
 }
@@ -139,10 +152,36 @@ function exitHandler(options, exitCode) {
   if (options.exit) process.exit();
 }
 
-const boldRed = (message) =>
-  `\u001b[1m\u001b[31m${message}\u001b[39m\u001b[22m`;
-const boldGreen = (message) =>
-  `\u001b[1m\u001b[32m${message}\u001b[39m\u001b[22m`;
+/**
+ * Captures groups from regex and executes RegEx.exec() function on all.
+ * 
+ * @param {regex} regex - Regular Expression to execute on.
+ * @param {string} template - HTML Template in string.
+ * @return {object} sandbox
+ * sandbox.matches - all matches of regex
+ * sandbox.input - input string
+ */
+const execRegexOnAll = (regex, template) => {
+  /** allMatches holds all the results of RegExp.exec() */
+  const allMatches = [];
+  let match = regex.exec(template); 
+  if (!match) {
+    return {matches: [], input: template};
+  }
+
+  const {input} = match; 
+
+  while (match !== null) {
+    delete match.input;
+    allMatches.push(match);
+    match = regex.exec(template);
+  }
+
+  return {matches: allMatches, input};
+};
+
+const boldRed = (message) => `\u001b[1m\u001b[31m${message}\u001b[39m\u001b[22m`;
+const boldGreen = (message) => `\u001b[1m\u001b[32m${message}\u001b[39m\u001b[22m`;
 const grey = (message) => `\u001b[90m${message}\u001b[39m`;
 
 module.exports = {
@@ -153,6 +192,7 @@ module.exports = {
   createPathIfAbsent,
   copyFolderSync,
   exitHandler,
+  execRegexOnAll,
   boldGreen,
   boldRed,
   grey,